refactor: move layout code out of herosimple block
